ENH: Use DICOM subseries tag values in loaded node name
Subseries tag values were not used in loadable image names because some of them were too long and complicated. However, there are simple values that are better to be displayed as is.

Added subseriesTagsToEnumerateValues variable to select which tags should be displayed as indices (1, 2, ...) and which ones should be displayed as values.

Since seriesInstanceUID, imageOrientationPatient, and diffusionGradientOrientation values tend to be long and complicated, they are displayed indices, while all others are displayed as values.