Dicom scalar plugin remove localizer image (#5055)
* ENH: Detect localizer images in cone-beam CT series

Siemens Axiom Artis saves a localizer image mixed into the image series. The previously added heuristics does not detect it, as there are multiple candidate subseries.

The new (additional) heuristics detects localizer image by checking if there is a tag that splits the series into two parts: a single image and the rest of the images; and the subseries does not have warning but the full series does has warning.

* ENH: Use DICOM subseries tag values in loaded node name

Subseries tag values were not used in loadable image names because some of them were too long and complicated. However, there are simple values that are better to be displayed as is.

Added subseriesTagsToEnumerateValues variable to select which tags should be displayed as indices (1, 2, ...) and which ones should be displayed as values.

Since seriesInstanceUID, imageOrientationPatient, and diffusionGradientOrientation values tend to be long and complicated, they are displayed indices, while all others are displayed as values.