ENH: Improve DICOM browser advanced section layout
Table column widths are optimized (loadable name column is interactive, as it can be long; plugin name is short so the column is set to content width; warning column uses the remaining space).
Plugin list is hidden by default and can be displayed using an expand button (double-arrow) to make more space for the loadable table.