ENH: Utilize color table node for segmentation import/export
When importing a segmentation node from file, users can now specify an existing color table node to create the segments with the expected color and name.
The same (or different) node can be used when exporting the segmentation to labelmap to ensure that the label values of the output match the input.

Users can now also export segmentations directly to NRRD/NIFTI from the "Export to files" section of the segmentation module. The same color node can be used to ensure that the correct expected label values are used during export.

re #5044