ENH: Display warnings and errors in the Save Data dialog.
This commit updates the Save Data dialog to include a new column
displaying warnings and errors associated with each row/file. It
addresses issue #4968.

To support this, the vtkMRMLStorageNode class has been updated to
include a new vtkMRMLMessageCollection object, to which
vtkMRMLStorageNode methods and the Save Data dialog can add messages
for the user.  Each warning and error is displayed in the Save Data
dialog as an icon in a new last column.  The tooltip for each icon is
the associated, detailed user message.

Co-authored-by: Andras Lasso <lasso@queensu.ca>
Co-authored-by: Steve Pieper <pieper@isomics.com>