BUG: Make sure subject hierarchy context menu is always up-to-date
If a plugin was registered later than the tree view created, then the context menus from the new plugin were not available. With this change, the context menu for each tree view is re-generated if a new plugin was registered since the last context menu generation.

Re https://github.com/SlicerMorph/SlicerMorph/issues/71