BUG: Fix module finder window showing up in taskbar preview on Windows
On Windows, dialog boxes that are just hidden but not deleted appear in taskbar preview (hover over the application icon in the taskbar, wait for the small preview window to appear, then hover over the preview window), which is quite confusing.

Deleting and recreating the module finder for each module switch would solve the problem, but rebuilding the view can take noticeable amount of time, so it is better to create the window only once.

Setting window style to Qt::Tool excludes the window from taskbar preview, but it also interferes with style and focus behavior of the window. Therefore we set this style, but only while the finder is hidden.