ENH: Improve display performance if TransformModifiedEvent is invoked
When a TransformModified event was invoked on any node in the scene, all model display pipelines were updated. This caused a noticeable drop in performance when there were a large number of nodes in the scene.

Fixed by not setting UpdateFromMRMLRequested to true when a transform was updated, and instead calling  OnMRMLDisplayableModelNodeModifiedEvent.

The performance of OnMRMLDisplayableModelNodeModifiedEvent was also improved, as it was not necessary to call UpdateClipSlicesFromMRML, and SetUpdateFromMRMLRequested functions when only a display node was modified.