COMP: add rpath to QWebEngineProcess
Other parts of the Qt installation have existing rpath
info in the mach-o file so they require the replacement
form of the install_name_tool command.

But QWebEngineProcess does not have an rpath, so
as described in the man page, this creates an
error condition so the other replacements are not
made.

This change uses the -add_rpath command command instead.

Fixes #5022
Fixes #4999