BUG: Make slice offset slider range span between slice centerpoints
Slice slider offset covered are between the two edges of the image (side of the voxels). This was a problem because when the slider was manually moved all the way to one side, then the slice position was not centered anymore but was set at the boundary between two slices.

Changed range to start at the first voxel center and end at the last voxel center. This way the slice position set by the slider remains centered on the slice.

Fixes #3346.