BUG: Fix crash when loading closed surface segmentations
When loading a segmentation from closed surface, the vtkMRMLSegmentationStorageNode will cause a crash if one of the blocks in the seg.vtm is unable to be loaded. Fixed by checking that the block could be read before trying to access it, and logging an error message if it could not.