BUG: Fix batch processing issues with subject hierarchy
The state of the subject hierarchy tree was not properly updated after batch processing. For example, after loading DICOM series using batch processing (while there was an active SH model, which is so if the user switches to the DICOM module) the series items were not under the corresponding study, but top-level.
The error was fixed by calling the proper update function for all the SH items.