BUG: Allow resetting color in terminology selector when non-default color is chosen
When a non-default color was automatically generated for a segment and user wanted to reset it to the default then user had to select another segment and then select the segment again, because the "Reset to default color" button was disabled.

Fixed by making "Reset to default color" button enabled when the default color is not the same as the current color.