BUG: Prevent cropping of segments imported from models
If the reference geometry conversion parameter had not been set, performing a closed surface to binary labelmap conversion would create the reference geometry if it didn't already exist, setting it to the geometry of the first imported model. This caused all subsequent models to be cropped to the same extent as the first one when exported or saved as labelmap.

Fixed by not setting the reference image geometry from vtkClosedSurfaceToBinaryLabelmapConversionRule, and by calculating extent from the union of segments in vtkSegmentation::DetermineCommonLabelmapGeometry if the reference geometry was not specified, and the extent mode was EXTENT_REFERENCE_GEOMETRY.

fixes #5093