ENH: Add option to toggle the markup measurements included in the description
Adds the MeasurementsInDescription attribute to vtkMRMLMarkupsNode, that controls the markups added to the description by WriteMeasurementsToDescription.
If MeasurementsInDescription is empty, then the description will be empty, and the measurement label in the 3D/2D views will not be displayed.

MeasurementsInDescription can be controlled from the markups subject hierarchy plugin. Each measurement adds a checkbox to the context menu to enable/disable display.