ENH: Improve error logging in subject hierarchy node
In many functions in the subject hierarchy node if the item was not found for any reason, the same message was logged "Failed to find subject hierarchy item by ID N", and it was impossible to know if it was caused by an internal problem in subject hierarchy, or because of the application that called the function with invalid item ID. This change makes sure that if an invalid item ID is given to these functions, then it's logged explicitly, so the developer can differentiate between application and internal SH bugs.

Also fixed a few copy-paste errors returning wrong values after errors.