BUG: Fix duplicate python logging messages in the log
Python INFO, WARNING, and ERROR messages generated by the python logging module created duplicate entries as seen in the Slicer log file and in the Error Log Dialog.  This commit removes python logging messages from showing in the python console as they are already included in the Slicer log and should be viewed there.