ENH: New Scalars display widget from Models module Scalars section
In order to use the scalar coloring features for other data types, it was necessary to create a new reusable widget for the related settings. The widgets from the Scalars section of the Models module were outsourced to a new Libs/MRML/Widgets/qMRMLScalarsDisplayWidget class. It seems to work perfectly within the Models module.