BUG: Fix writing of if window width&level during DICOM export
Due to a bug, the mandatory window width and window level fields were not written to the exported DICOM file.

Also, the computed values were not ideal, as the window width/level corresponded to the full scalar range of the volume.
Kept this as the current behavior in CreateDICOMSeries module, but also added the option to specify a custom width/level.
Using this new API, DICOMExportScalarVolume plugin now writes the current display window width/level into the exported DICOM files.