BUG: fix observer manager events and priorites
The method gives the option of providing a list of priorities
with one priority per event in two lists.

The check logic was incorrect however, because in the case
where no priorities had been provide (null list pointer) the
'if' condition was triggered and the warning macro would
indirect through the null priorities pointer.

Now the two checks have been separated and each has a custom
warning message so that pointers are only used when they are
known to be non-null.