BUG: Fix markup selection with VTK9 updating tolerance
Following the revamp of the z-buffer computation and coincident topology
resolution integrated in VTK9 (See [1][2]), the z-buffer values are
now different. This commit accounts for this change by increasing the
world tolerance associated with small spheroid objects (such as glyphs)
and by setting a tiny tolerance for normalized display coordinates.
These tolerances are used to configure vtkSelectVisiblePoints filter
used to determine visibility of markups used to enable/disable
selectability.

[1] update and improve the coincident rendering code
    https://github.com/Kitware/VTK/commit/3052137e6621684f9184b293b82c677e976c2f89

[2] Add support for camera based shift scale methods
https://github.com/Kitware/VTK/commit/8f3aef771a25b81d89ae591ef26aca5162c8c6b2

Fixes #5250

Co-authored-by: Andras Lasso <lasso@queensu.ca>