COMP: Rename external project install dir from <proj>-install to <proj>-i
This commit rename the install directory to be consistent with the build
directory naming convention introduced in previous commit.

Renaming the install directory is done for sake of consistency and not
for directly addressing "command line too long" type of errors. Indeed,
since all headers are installed in the same directory, passing such include
directory to the compiler does not trigger such issue.