BUG: Always add excluded nodes to subject hierarchy
The purpose of excluded nodes is hide them in the UI. However, excluded attributes can change, so we need to have these nodes in subject hierarchy so that these changes can be monitored. The proxy model can decide if they should be shown in the UI or not.

Note that the HideFromEditors flag is not considered to work dynamically, meaning that we don't expect changes on the UI when we set it after adding the node to the scene, so not adding hidden nodes to SH should not cause issues. However, adding them like now the excluded nodes could slow down the application due to the large number of hidden nodes in every scene (color nodes etc.)