BUG: Fix markup position jumping when using interaction handles
Previous implementation of interaction handle translation/rotation tried to align the position of the handle to the current mouse position, which had the issue that the markup handle would "jump" if the user didn't click exactly on the center of the handle.
This is fixed by transforming the markup handles based on the difference between the current and previous mouse positions.

fixes #5257