ENH: Allow showing nodes by drag-and-dropping into views from subject hierarchy tree
A node can be shown in a view by drag-and-dropping it from the subject hierarchy into a view.
If a volume is dragged to a 3D view then volume rendering is displayed.
To show a node in a single view only, hide the node (click on the eye icon in subject hierarchy) and then drag-and-drop it into a view.
To restore visibility of a node in all views, choose "Show in all views" in subject hierarchy item visibility.