ENH: Improve Endoscopy module with markups curve input and minor fixes
Previously, if Endoscopy module got a markups curve node as input, it recreated the curve from control points. This resulted in slightly different path as the original curve.
Now the module takes the curve points provided by the markups node (and resamples them at equal distances).

Also added an output model node selector for storing the path (and the path node maintains references to the transform and cursor nodes) so that the path can be regenerated without creating new nodes all the time.

Also fixed transform computation (it was not orthogonalized properly) and made the cursor transparent when looked from inside.