BUG: Fix selection in subject hierarchy tree to jump back to previous selection
When user clicked on an item in subject hierarchy after drag-and-dropping another item to viewers (to display), the item selection sometimes reverted to the previously selected item.

The problem was that when the drag operation did not end up with dropping inside the subject hierarchy tree, which resulted in DraggedSubjectHierarchyItems list not being cleared out and so on the next update operation, this list was used to restore item selection.

Fixed by restoring item selection from DraggedSubjectHierarchyItems only if items are dropped _within_ the widget (because this is when reparenting happens and when item selection is lost).