BUG: Prevent clipping of volumes when hardening non-linear transform
When non-linear transform was hardened on a volume then only the 8 corners of the volume were used for determining bounding box and therefore parts of the volumes that bulged out of the original bounding box got clipped.

Fixed the issue by improving vtkMRMLVolumeNode::GetBoundsInternal to compute transformed by extent by sampling the volume (grid of 8x8x8 over the full extent), and using these bounds in vtkMRMLVolumeNode::ApplyNonLinearTransform to determine bounds of the output volume.