BUG: Fix error message display in Extensions manager
"There is no application set up to open URL data:text..." error was displayed when qSlicerExtensionsInstallWidgetPrivate::setFailurePage method was called instead of showing an error page.

The problem was that QWebEngineView::setHtml passes the html content as a special URL (data:...). qSlicerWebWidget did not recognize this as an internal URL, therefore it tried to open it in an external application. However, external applications do not recognize this special URL.

Also improved the error message on MacOS (tell the user to install the application).

See #5022