ENH: Update CTK to latest master
Contains fixes and improvements:
* ENH: Make DICOM database relocatable
* ENH: add command and history properties to ctkConsole
* ENH: Remove warning message when nullptr is used as vtkObject pointer event data

Details:

Revision: 2902545028332492d77abed9223a3390f0e5e33e
Author: Andras Lasso <lasso@queensu.ca>
Date: 2020-12-30 11:49:24 PM
Message:
ENH: Make DICOM database relocatable

Previously full path was stored to DICOM files that were copied into the DICOM database. This made it impossible to move the database into any other location or store in on removable storage (such as a USB stick, which may be mapped with different path on different computers).

Now files that are copied (not linked) to the DICOM database are stored with path relative to the DICOM database folder (where the .sql files are located).

Also, DICOM database location can now be specified with relative path in application settings. This allows the database to be moved along with the application. For example, to implement a DICOM viewer that opens DICOM database in a folder relative to the application's location (e.g., use the application as a viewer on a DICOM CD).

Revision: 1f577aadf0775dfe700b5b45dd9690fb4af31cf9
Author: Steve Pieper <pieper@isomics.com>
Date: 2020-12-31 1:21:54 PM
Message:
Merge pull request #945 from pieper/console-enh

ENH: add command and history properties to ctkConsole

Revision: 9413e18ec38ba0f5e0bc66ef1254935f2e0cc4b6
Author: Steve Pieper <pieper@isomics.com>
Date: 2020-12-30 6:25:57 PM
Message:
ENH: add command and history properties to ctkConsole

Previously, the API allowed execution of commands but not
setting the state of the command buffer to be executed when the
user hit enter.

Also the command history always started empty.

This change adds read/write properties to fix this.

Now the command buffer can be set, e.g. by a hot key or other
custom code and it will be editable and the user can add more text,
user the completer, etc.

Also the command history can be controlled by the appication so
that it can save and restore previous sessions as a convenience
to the user.

Revision: ac0cadfe10b200b2ae9a7e8d9c0b09936fb7f254
Author: Steve Pieper <pieper@isomics.com>
Date: 2020-12-08 4:34:27 PM
Message:
Merge pull request #944 from lassoan/remove-warning

ENH: Remove warning message when nullptr is used as vtkObject pointer event data

Revision: 35a169e160b9b2556707ec6cea0a7f862cd47378
Author: Andras Lasso <lasso@queensu.ca>
Date: 2020-12-08 3:45:28 PM
Message:
ENH: Remove warning message when nullptr is used as vtkObject pointer event data

ctkVTKConnection logged a warning message each time an event was received that vtkObject* as event data and the pointer was nullptr.

Always treating nullptr value as an error is too limiting, because it is often useful that an additional pointer can be supplied but not mandatory.

This change removes the warning message from ctkVTKConnection and makes it the event receiver's responsibility to log warning/error if in that particular case nullptr is not expected.
Since ctkVTKConnection just logged a warning message but already left nullptr values through, it is guaranteed that change does not alter any existing software behavior
(other than not displaying the warning).

This change is useful in Slicer because DisplayModified events typically contain the display node pointer as eventdata but if multiple DisplayModified events are lumped together
(due to batch update with Start/EndModify) then there is no display node that could be attached to the event.