BUG: Fix Qt printer support for Slicer installations
For installed versions of Slicer, calling QPrinterInfo::availablePrinterNames() would return an empty list, while built versions would have the list of printers populated as normal.
The lack of available printers made it impossible to set the output of QPrinter to NativeFormat.

By bundling the printsupport plugin with the installer, the available printers can be found.