ENH: Enable automatic module logic registration
This enables automatic set and unset of module logics
in vtkSlicerApplicationLogic when modules are loaded/unloaded byt the
qSlicerModuleFactoryManager.

On initialization, modules initialize subject hierarchy plugins. Those
subject hierarchy plugins require the module logic to be registered
beforehand. Here, we unregister the module logic if the title (required)
is not set.