BUG: Fix character encoding of command-line arguments
Use QString's built-in toWCharArray converter because that is guaranteed to convert correctly to wchar.
Python's Py_DecodeLocale converter only works well for the QString's UTF-8 output if Python determines
that locale is UTF-8 (which should be the case most of the time, but there might be special environments
where it is not).