DOC: Remove subcategories in User manual's modules section
A page can only appear at one place in the navigation tree (otherwise opening a page could jump to a different part of the tree).
Since it is hard to find a module if it is hidden in a subcategory, it is better to have a simple list only.

In the future, when many more modules will be documented, we may need to find a way to improve navigation,
but this may take a long time and in the meantime documentation generation tools might improve.