BUG: Fix vtkSegmentationConverter::GetCheapestPath if cost >= 10000000
For some Segmentation conversion methods vtkSegmentationConverterRule::GetConversionInfiniteCost() (10000000) is used to represent a last resort conversion path that should not be used unless it is the only one that can be found.
GetCheapestPath initialized the cheapestPathCost using 10000000. This meant that the cheapest path would never be found if its cost was >= 10000000.

Fixed by initializing cheapestPathCost to VTK_UNSIGNED_INT_MAX.

See SlicerRt/SlicerRT#177