BUG: Remove node modified calls from background thread
Problem:
itkMRMLIDImageIO is used in a background thread when running CLI modules or Simple Filters module and it called node Modified events when it finished updating a node in the scene.
Since node modified events can trigger a cascade of method calls, invoking events from a different thread can easily lead to application crash or data correction.

Solution:
Replaced the direct modified event invocation by a thread-safe modified request call.
RequestModified method was already implemented in vtkSlicerApplicationLogic, but this class was not directly visible to MRML core classes.
To allow calling the method without introducing explicit dependency between these classes, a callback command was added to vtkEventBroker.
The application connects this callback command to the application logic when it creates the application logic object.

Fixes the crash reported here:
https://discourse.slicer.org/t/reproducible-slicer-crash-when-pushing-to-some-volume-nodes-using-sitkutils-pushvolumetoslicer/15723/3