BUG: Fix incorrect/missing volume node copy content macros
In vtkMRMLVolumeNode, vtkMRMLVectorVolumeDisplayNode was used as the class name in vtkMRMLCopyContentMacro, rather than vtkMRMLVolumeNode. Similar issue in vtkMRMLDiffusionImageVolumeNode.
Fixed by updating the macros with the correct class name.

vtkMRMLVectorVolumeNode was missing a vtkMRMLCopyContentDefaultMacro() so HasCopyContent() always returned false.
This prevented vector volume nodes from being displayed in the list of proxy nodes in the Sequences module.
Fixed by adding the missing macro.