BUG: Show default layout if the scene uses an undefined layout id
Modules can define custom view layouts. When a scene that used custom view layout was loaded in a Slicer instance that did not have that layout defined,
then the view area appeared blank.

This commit fixes the issue by showing default view layout when an unknown layout ID is set in the layout node.