ENH: Improve "not compatible with current Slicer version" message
When extension manager was too slow to respond (and so queries timed out) or an extension was not yet built for a particular Slicer version then the Extensions manager restore tab showed the message "not compatible with current Slicer version". This made user think that the extension is permanently not available for the current Slicer version, while in most cases it is a transient error.

Update error message to just say "not found for current Slicer version" to not suggest that it is permanent.