ENH: Add Markups ROI node
Adds an ROI markup type with the vtkMRMLMarkupsROINode class.
ROI dimensions can be changed by clicking on the interactive handles on the faces and corners of the ROI.

The ROIType property of vtkMRMLMarkupsROINode controls how the ROI is defined:
- BOX: Users can place two control points to define the size of the ROI. Once the two points are placed, they are removed.
- BOUNDING_BOX: ROI encompasses all of its control points.

This commit also adds support for the markups ROI node to be used in the Volume Rendering module.