BUG: Ensure temporaryPath is writable
To ensure that CLI modules, sample data downloads, etc. work it is necessary to have a writable temporary directory.
In some cases the path specified in application settings does not point to a valid, writable location anymore.

To fix this, the application checks at startup that the temporary path is writeable. If the check fails then the path is reset to the default.

see #4652