ENH: Improve default axis-aligned display of single-slice volume
When "Reset view orientation on show" was enabled then single-slice volumes fully appeared in all view (view normal aligned with slice normal), with different rotations.
This resulted in redundant and slightly confusing views.

Fixed the problem by forcing slice view normal to be aligned with the volume slice normal only if a volume is requested to be shown in a single view.
If a volume is shown in multiple views then each slice view is aligned with the nearest volume axis.