ENH: Ensure that remote cache folder is writable
macOS started to manage application temporary and cache folders more actively, i.e., remove them after certain time period (or at system restart or application exit) to improve security.

Cache folder by default was initialized to be a subfolder of temporary folder. However, when the system removed the temporary folder, the cache folder has become invalid.

To fix this, implemented an automatic mechanism to check the cache folder at startup and reset it to a system-recommended location if the folder is not writable.

See related user error report: https://discourse.slicer.org/t/slicer-4-11-on-m1-not-loading-mrb-file/16228/7