BUG: Fix display of CPU volume rendering of volumes with small voxel size
The CPU volume renderer in VTK does not take into account the user matrix (that we use to position, orient, and scale the volume). It would be hard to change this, because it is a complex mechanism and the CPU renderer is not developed anymore in VTK. However, we can work around this by passing just the scaling to the volume renderer, and use the user matrix only for positioning and scaling.