ENH: Use flow layout in Sample Data module
The current three-column box layout resizes the module panel if it was narrower than what could fit three thumbnails, which was an inconvenience especially on smaller monitors or smaller window sizes, because the user had to manually resize the panel after switching to another module if they wanted to see the viewers in the usual size.
By using flow layout the thumbnails naturally fit the available horizontal space.