BUG: Fix circular markups measurement update calls
The custom modified event InputDataModifiedEvent has been introduced in vtkMRMLMeasurement that is invoked only when such members are modified that should trigger update of the measurement. By observing this event for updating the measurements, circular calls are prevented.

A concrete effect is that curvature computation and scalar display update is fast again.

Fixes https://github.com/Slicer/Slicer/issues/5486