BUG: Fix new segments appearing as hidden
Segment selector widget in qMRMLSliceControllerWidget made segments hidden when the segmentation display node was updated earlier than the segmentation node.
The problem was that qMRMLSliceControllerWidget made all segments that were not in the segment selector widget hidden, and segments were not yet in the segment selector widget when they have just been added but node modified event was not yet invoked.

Fixed by not modifying visibility of those segments that are not yet in the segment selector widget.

Fixes #5533