ENH: Use parallel transport for more robust curve coordinate system computation
Previously, "consistent" variant of Frenet-Serret method (in SplineDrivenImageSlicer VTK remote package) was used in GetCurvePointToWorldTransformAtPointIndex to get coordinate frame along the curve. However, this implementation had random initial normal direction and it was not robust when points were too close to each other.

This commit switches to a parallel transport method, which is more robust and predictable. Initial direction is set so that +X direction is preferred as normal direction (if curve tangent is exactly +X then +Y direction is preferred as binormal direction).

Fixes #5559