BUG: Fix crash after deleting a transform that had a 3D widget
This steps resulted in a crash: create new transform, show 3D widget, delete transform, click in a 3D view, hit any key.

The problem was that a keyboard interaction observer remained active after the transform widget was deleted that called method of the deleted widget.

Fixed by disabling the widget before deleting it. This removes all the observers properly.