ENH: Improve sequence browser toolbar usability
Display sequence toolbar when sequence is loaded from scene. Before this, sequence toolbar was displayed when an individual sequence was loaded; but was not displayed when the sequence was loaded as part of a scene.

Display sequence browser toolbar in a separate row. The toolbar is usually already quite busy and if the large sequence browser toolbar is added there then there is no space for many toolbar buttons and they get hidden behind hardly visible '>>' button submenus. The user can move the sequence browser toolbar manually and this manual position is stored/restored along with other window and toolbar positions on application restart.