ENH: Explicitly set AppUserModelID on Windows for improved taskbar appearance
AppUserModelID is used for grouping and pinning applications on the taskbar on Windows. Since AppUserModelID of the launcher (Slicer.exe) and the real application (SlicerApp-real.exe) were different, they appear as separate items (if console was displayed) and pinning the Slicer application created a broken shortcut (because it ran SlicerApp-real.exe without a launcher).

Explicitly setting AppUserModelID in SlicerApp-real.exe and setting the same AppUserModelID in the shortcut that launches Slicer solves both issues.

As an additional benefit effect, the application can be referred to using this global AppUserModelID identifier (which is generated from organization, application name, and full version - for example: NA-MIC.Slicer.4.13.0-2021-12-01). The application can be launched using `shell:AppsFolder\NA-MIC.Slicer.4.13.0-2021-12-01` (or from the terminal: `explorer shell:NA-MIC.Slicer.4.13.0-2021-12-01`).

Co-authored-by: Andras Lasso <lasso@queensu.ca>