BUG: Fix invalid resolution of stray node references in saved scenes
When a scene contained stray node references (referenced node ID set without a corresponding node in the scene) then that node reference was preserved when the scene is imported.
This node reference could then connect the node to unrelated nodes that existed in the scene before or added later.

When switching between CPU and GPU VR, VR display nodes were removed from the scene, but node references were not removed (because this was all done in batch processing mode).
As a result, when a new VR display node was added, it got assigned to both to the new volume and to the old volume (that had still kept a reference to the same node ID).

Fixed the issue by immediately removing node references, even in batch processing mode. This may fix node reference issues elsewhere, too.

Also made more robust the function that returns non-volume-rendering display node (vtkMRMLVolumeNode::GetVolumeDisplayNode). It returns the first non-volume-rendering display node, not just the very first node. This is important because volume rendering display node may be the first display node when a node is loaded while volume rendering module is active.