BUG: Prevent flood of terminology errors in the log if using an older scene
The commit 5ae5564362b40e43730a289e1fa0531feee8ca73 on 2020.03.17 changed all the terminology category codes in the default terminology contexts. As a result, if a scene saved before this date containing terminology information is loaded in a version newer than that, lots of errors are reported about the category not found in the specified terminology category.

The errors logged are pairs of entries looking like this:
"GetCategoryInTerminology: Failed to find category 'Anatomical Structure' in terminology 'Segmentation category and type - 3D Slicer General Anatomy list'
DeserializeTerminologyEntry: Failed to get terminology category"