BUG: Fixing missing volume centering in 3D view
When in Data node "Reset field of view on show"  and "Show volume rendering" were enabled for a volume, clicking the eye icon to show the volume in all views, made the volume only centered in slice views, but not in 3D views.

This commit fixes this behavior by centering the volume in all the 2D and 3D views where the volume is displayed.