ENH: Add code of conduct document
While reviewing project metrics for a grant application we realized that although we have a common agreement about how to conduct operations in the Slicer community, we do not have a formalized document to describe it.

The Contributor Covenant is a short and clear description that captures well how we operate, so we propose to adopt this as code of conduct document.

The only customization that is made is the contact information: "...reported to the community leaders listed
[here](https://github.com/Slicer/Slicer/blob/master/CONTRIBUTING.md#benevolent-dictators-for-life)
via private message on the [Slicer forum](https://discourse.slicer.org)."