ENH: Add create/remove representation options to segmentations subject hierarchy menu
It is quite frequently needed and not easy for users to find how to create closed surface representation for a segmentation node.

This commit adds "create" and "remove" actions for "closed surface" and "binary labelmap" representations.
Only the non-master representation is offered to be created or deleted, therefore typically only one action appears in the menu.
For example, if binary labelmap is the master representation then either "Create closed surface representation"
or "Remove closed surface representation" action is shown.