ENH: Improve performance of markups visibility detection
Previously, each ControlPointsPipeline3D had its own vtkSelectVisiblePoints filter to determine if the markups were occluded by opaque geometry in the render window. The filters would all download the ZBuffer data separately, resulting in a performance loss.

This commit adds the vtkFastSelectVisiblePoints filter which exposes the option to specify the Z-buffer pointer as a vtkFloatArray. As a result, we now only get the contents of the Z-buffer once, and re-use the same data for each control point pipeline.